[ENG-11756] Isolate swizzle state
