+++ SDKTest/NeuroIDClass/MultiAppFlowTests.swift
@@ -9,6 +9,7 @@ import XCTest
 
 @testable import NeuroID
 
+@MainActor
 class MultiAppFlowTests: XCTestCase {
     let clientKey = "key_test_0OMmplsawAp2CQfWrytWA3wL"
