+++ Source/NeuroID/NeuroIDClass/Extensions/NIDAdvancedDevice.swift
@@ -8,6 +8,7 @@
 import Foundation
 
 extension NeuroIDCore {
+    @MainActor
     func start(
         _ advancedDeviceSignals: Bool,
         completion: @escaping (Bool) -> Void = { _ in }
@@ -23,6 +24,7 @@ extension NeuroIDCore {
         }
     }
 
+    @MainActor
     func startSession(
         _ sessionID: String? = nil,
         _ advancedDeviceSignals: Bool,
