+++ Source/NeuroID/NeuroIDClass/Extensions/NIDSession.swift
@@ -59,6 +59,7 @@ extension NeuroIDCore {
        throughout the rest of the session
       i.e. start/startSession/startAppFlow -> startAppFlow("site2") -> stop/stopSession
      */
+    @MainActor
     func startAppFlow(
         siteID: String,
         sessionID: String? = nil,
@@ -251,6 +252,7 @@ extension NeuroIDCore {
      - Will move queued events into main queue
      - Will make call to check/capture ADV event
      */
+    @MainActor
     func setupSession(
         siteID: String?,
         customFunctionality: @escaping () -> Void = {},
@@ -266,7 +268,7 @@ extension NeuroIDCore {
         self.setupListeners()
 
         self.createSession()
-        self.swizzle()
+        self.uiRuntime.swizzle()
 
         // custom functionality = the different timer starts (start vs. startSession)
         //  this will be refactored once we bring start/startSession in alignment
@@ -280,6 +282,7 @@ extension NeuroIDCore {
     }
 
     // Internal implementation that allows a siteID
+    @MainActor
     func start(
         siteID: String?,
         completion: @escaping (Bool) -> Void = { _ in }
@@ -313,6 +316,7 @@ extension NeuroIDCore {
     }
 
     // Internal implementation that allows a siteID
+    @MainActor
     func startSession(
         siteID: String?,
         sessionID: String? = nil,
