+++ Source/NeuroID/UIRuntime.swift
@@ -3,10 +3,12 @@
 //  NeuroID
 //
 
-import Foundation
+import UIKit
 
 @MainActor
 final class UIRuntime {
+
+    var didSwizzle: Bool = false
 
     var sceneCaptureRegistrationsBySceneID: [String: AnyObject] = [:]
     var sceneCaptureLastKnownStateBySceneID: [String: Bool] = [:]
@@ -20,5 +22,17 @@ final class UIRuntime {
         sceneCaptureRegistrationsBySceneID.removeAll()
         sceneCaptureLastKnownStateBySceneID.removeAll()
         screenCaptureLastKnownState = nil
+    }
+
+    func swizzle() {
+        guard !didSwizzle else { return }
+
+        UIViewController.startSwizzling()
+        UITextField.startSwizzling()
+        UITextView.startSwizzling()
+        UINavigationController.swizzleNavigation()
+        UITableView.tableviewSwizzle()
+
+        didSwizzle = true
     }
 }
